+++ src/App.jsx
@@ -12,27 +12,13 @@ import Groups from "./pages/Groups";
 import Login from "./pages/Login";
 
 import NotFound from "./pages/NotFound";
-import ErrorHandler from "./components/auth/authorize/ErrorHandler";
-import axios from "axios";
-import Register from "./pages/Register";
-import { store } from "./redux/rootReducer";
 import PrivateRoute from "./components/routing/PrivateRoute";
 import PublicRoute from "./components/routing/PublicRoute";
+import axios from "axios";
 import { HELPER, STUDENT, TEACHER } from "./redux/state/stateActions";
 
 const BASE_PATH = "https://dss-course-work.herokuapp.com/api/v1";
 axios.defaults.baseURL = BASE_PATH;
-axios.interceptors.response.use(
-  (response) => {
-    return response;
-  },
-  (error) => {
-    return Promise.reject(
-      ErrorHandler(store, error)
-    );
-  }
-);
-axios.defaults.headers.common['Authorization'] = `Bearer ${JSON.parse(localStorage.getItem("auth"))?.user.token }`;
 
 function App() {
   return (
