+++ src/components/auth/Auth.jsx
@@ -5,7 +5,7 @@ import { Nav, Button } from "react-bootstrap";
 import { NavLink } from "react-router-dom";
 import { logouted } from "../../redux/auth/authReducer";
 
-function AuthPanel() {
+function Auth() {
   const auth = useSelector(selectAuth);
   const dispatch = useDispatch();
 
@@ -36,4 +36,4 @@ function AuthPanel() {
   }
 }
 
-export default AuthPanel;
+export default Auth;
