--- src/components/auth/login/login-form.less
@@ -1,12 +1,0 @@
-.login-form {
-    max-width: 500px;
-}
-
-@media (max-width: 500px) {
-    .login-form {
-        max-width:none;
-        min-width: 250px;
-        padding-left: 0%;
-        padding-right: 0%;
-    }
-}
