--- src/components/auth/register/register-form.less
@@ -1,12 +1,0 @@
-.register-form {
-    max-width: 500px;
-}
-
-@media (max-width: 500px) {
-    .register-form {
-        max-width:none;
-        min-width: 250px;
-        padding-left: 0%;
-        padding-right: 0%;
-    }
-}
