--- src/pages/Register.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { Container } from "react-bootstrap";
-import RegisterForm from "../components/auth/register/RegisterForm";
-
-function Register() {
-  return (
-    <Container>
-      <RegisterForm />
-    </Container>
-  );
-}
-
-export default Register;
